LayoutTests/imported/w3c/web-platform-tests/XMLHttpRequest/setrequestheader-header-allowed-expected.txt

CONSOLE MESSAGE: line 16: Refused to set unsafe header "User-Agent"
CONSOLE MESSAGE: line 17: Refused to set unsafe header "User-Agent"
CONSOLE MESSAGE: line 16: Refused to set unsafe header "Content-Transfer-Encoding"
CONSOLE MESSAGE: line 17: Refused to set unsafe header "Content-Transfer-Encoding"

FAIL XMLHttpRequest: setRequestHeader() - headers that are allowed (Authorization) assert_equals: expected "authorization," but got ""
FAIL XMLHttpRequest: setRequestHeader() - headers that are allowed (Pragma) assert_equals: expected "pragma," but got ""
FAIL XMLHttpRequest: setRequestHeader() - headers that are allowed (User-Agent) assert_equals: expected "user-agent," but got ""
FAIL XMLHttpRequest: setRequestHeader() - headers that are allowed (Content-Transfer-Encoding) assert_equals: expected "content-transfer-encoding," but got ""
FAIL XMLHttpRequest: setRequestHeader() - headers that are allowed (Content-Type) assert_equals: expected "content-type," but got ""
PASS XMLHttpRequest: setRequestHeader() - headers that are allowed (Overwrite) 
PASS XMLHttpRequest: setRequestHeader() - headers that are allowed (If) 
PASS XMLHttpRequest: setRequestHeader() - headers that are allowed (Status-URI) 
PASS XMLHttpRequest: setRequestHeader() - headers that are allowed (X-Pink-Unicorn) 
